Implementacion de product mapper

+++ lib/features/products/domain/entities/product.dart
@@ -1,10 +1,9 @@
-
- import 'package:teslo_shop/features/auth/domain/domain.dart';
+import 'package:teslo_shop/features/auth/domain/domain.dart';
 
 class Product {
   String id;
   String title;
-  int price;
+  double price;
   String description;
   String slug;
   int stock;
@@ -27,5 +26,4 @@ class Product {
     required this.images,
     required this.user,
   });
-
 }
